requirements.txt
azure-core==1.19.0
azure-data-tables==12.1.0
azure-storage-queue==12.1.6
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.7
cryptography==35.0.0
idna==3.2
isodate==0.6.0
lxml==4.6.3
msrest==0.6.21
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
pandas-datareader==0.10.0
pycparser==2.20
pyodbc==4.0.32
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.7
